Moved testable code from zimcheck to zimfilechecks
